CSS cleanup of Description section
