+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css"
-import Description from "./components/Description.js";
+import Description from "./components/Description/Description.jsx";
 import Input from "./components/Input.js";
 import StockGraph from "./components/StockGraph.jsx";
 import Output from "./components/Output.jsx";
