--- src/components/Description.css
@@ -1,26 +1,0 @@
-h1 {
-  font-size: 70px;
-  text-align: center;
-}
-
-p {
-  font-size: 25px;
-  line-height: 1.5;
-  margin-right: 200px;
-  margin-left: 200px;
-}
-
-.button-info-list {
-  font-size: 25px;
-  margin: 10px 250px 10px 250px;
-  line-height: 1.8;
-}
-
-.button-info {
-  color: green;
-  padding: 3px 3px 3px 3px;
-  border-style: solid;
-  border-width: 2px;
-  border-color: green;
-  border-radius: 5px;
-}
