+++ src/components/Input.css
@@ -13,3 +13,8 @@
   font-size: 25px;
   color: black;
 }
+
+button-style.hover {
+  background-color: lightgreen;
+  transition: .5s;
+}
